远程启动脚本;数据解析、返回接口

+++ make.sh
@@ -1,11 +1,20 @@
 #!/bin/bash
+set -e
 BIN_PATH="bin/test_app"
 clear
-rm ${BIN_PATH}
+
+if [ -f "$BIN_PATH" ]; then
+    rm ${BIN_PATH}
+fi
+
 cd build
 cmake ..
 make
 cd ..
 chmod 755 ${BIN_PATH}
 echo -e "build finish"
-./${BIN_PATH}
+# ./${BIN_PATH}
+ssh root@47.112.229.111 < script/kill.sh
+scp -r bin/test_app root@47.112.229.111:/root/
+clear
+ssh root@47.112.229.111 < script/start.sh
